@@ -17,6 +17,12 @@
 
 namespace grid_map {
 
+union Color
+{
+    unsigned long longColor;
+    float floatColor;
+};
+
 /*!
  * Gets the position of a cell specified by its index in the map frame.
  * @param[out] position the position of the center of the cell in the map frame.
@@ -142,7 +148,7 @@ void wrapIndexToRange(Index& index, const Size& bufferSize);
  * @param[in/out] index the index that will be wrapped into the valid region of the buffer.
  * @param[in] bufferSize the size of the buffer.
  */
-void wrapIndexToRange(int& index, const int& bufferSize);
+void wrapIndexToRange(int& index, int bufferSize);
 
 /*!
  * Bound (cuts off) the position to lie inside the map.
@@ -288,25 +294,6 @@ size_t getLinearIndexFromIndex(const Index& index, const Size& bufferSize, const
  * @return the regular 2d index.
  */
 Index getIndexFromLinearIndex(const size_t linearIndex, const Size& bufferSize, const bool rowMajor = false);
-
-/*!
- * Generates a list of indices for a region in the map.
- * @param regionIndex the region top-left index.
- * @param regionSize the region size.
- * @param indices the list of indices of the region.
- */
-void getIndicesForRegion(const Index& regionIndex, const Size& regionSize,
-                         std::vector<Index> indices);
-
-/*!
- * Generates a list of indices for multiple regions in the map.
- * This method makes sure every index is only once contained in the list.
- * @param regionIndeces the regions' top-left index.
- * @param regionSizes the regions' sizes.
- * @param indices the list of indices of the regions.
- */
-void getIndicesForRegions(const std::vector<Index>& regionIndeces, const Size& regionSizes,
-                          std::vector<Index> indices);
 
 /*!
  * Transforms an int color value (concatenated RGB values) to an int color vector (RGB from 0-255).
