@@ -40,7 +40,8 @@ namespace grid_map {
 		  for(int c = 0; c < A.cols(); c++)
 			{
 		    if (std::isnan(A(r,c))) {
-		      ASSERT_TRUE(std::isnan(B(r,c)));
+		      ASSERT_TRUE(std::isnan(B(r,c))) << message << "\nNaN check failed at (" << r << "," << c << ")\n"
+                                  << "\nMatrix A:\n" << A << "\nand matrix B\n" << B;
 		    } else {
 			    ASSERT_EQ(A(r,c),B(r,c)) << message << "\nEquality comparison failed at (" << r << "," << c << ")\n"
                                        << "\nMatrix A:\n" << A << "\nand matrix B\n" << B;
@@ -63,7 +64,8 @@ namespace grid_map {
 		  for(int c = 0; c < A.cols(); c++)
 			{
 		    if (std::isnan(A(r,c))) {
-		      ASSERT_TRUE(std::isnan(B(r,c)));
+		      ASSERT_TRUE(std::isnan(B(r,c))) << message << "\nNaN check failed at (" << r << "," << c << ")\n"
+                                  << "\nMatrix A:\n" << A << "\nand matrix B\n" << B;
 		    } else {
 			    ASSERT_NEAR(A(r,c),B(r,c),tolerance) << message << "\nTolerance comparison failed at (" << r << "," << c << ")\n"
 												   << "\nMatrix A:\n" << A << "\nand matrix B\n" << B;
@@ -83,7 +85,8 @@ namespace grid_map {
           for(int c = 0; c < A.cols(); c++)
             {
               if (std::isnan(A(r,c))) {
-                EXPECT_TRUE(std::isnan(B(r,c)));
+                EXPECT_TRUE(std::isnan(B(r,c))) << message << "\nNaN check failed at (" << r << "," << c << ")\n"
+                            << "\nMatrix A:\n" << A << "\nand matrix B\n" << B;
               } else {
                 EXPECT_NEAR(A(r,c),B(r,c),tolerance) << message << "\nTolerance comparison failed at (" << r << "," << c << ")\n"
                              << "\nMatrix A:\n" << A << "\nand matrix B\n" << B;
