@@ -57,9 +57,21 @@ inline Eigen::Matrix2i getBufferOrderToMapFrameTransformation()
   return -Eigen::Matrix2i::Identity();
 }
 
+inline Vector transformBufferOrderToMapFrame(const Index& index) {
+  return {-index[0], -index[1]};
+}
+
 inline Eigen::Matrix2i getMapFrameToBufferOrderTransformation()
 {
   return getBufferOrderToMapFrameTransformation().transpose();
+}
+
+inline Index transformMapFrameToBufferOrder(const Vector& vector) {
+  return {-vector[0], -vector[1]};
+}
+
+inline Index transformMapFrameToBufferOrder(const Eigen::Vector2i& vector) {
+  return {-vector[0], -vector[1]};
 }
 
 inline bool checkIfStartIndexAtDefaultPosition(const Index& bufferStartIndex)
@@ -74,7 +86,7 @@ inline Vector getIndexVectorFromIndex(
 {
   Index unwrappedIndex;
   unwrappedIndex = getIndexFromBufferIndex(index, bufferSize, bufferStartIndex);
-  return (getBufferOrderToMapFrameTransformation() * unwrappedIndex.matrix()).cast<double>();
+  return transformBufferOrderToMapFrame(unwrappedIndex);
 }
 
 inline Index getIndexFromIndexVector(
@@ -82,7 +94,7 @@ inline Index getIndexFromIndexVector(
     const Size& bufferSize,
     const Index& bufferStartIndex)
 {
-  Index index = (getMapFrameToBufferOrderTransformation() * indexVector.cast<int>()).array();
+  Index index = transformMapFrameToBufferOrder(indexVector);
   return getBufferIndexFromIndex(index, bufferSize, bufferStartIndex);
 }
 
@@ -126,8 +138,7 @@ bool getIndexFromPosition(Index& index,
   getVectorToOrigin(offset, mapLength);
   Vector indexVector = ((position - offset - mapPosition).array() / resolution).matrix();
   index = getIndexFromIndexVector(indexVector, bufferSize, bufferStartIndex);
-  if (!checkIfPositionWithinMap(position, mapLength, mapPosition)) return false;
-  return true;
+  return checkIfPositionWithinMap(position, mapLength, mapPosition) && checkIfIndexInRange(index, bufferSize);
 }
 
 bool checkIfPositionWithinMap(const Position& position,
@@ -165,7 +176,7 @@ bool getIndexShiftFromPositionShift(Index& indexShift,
     indexShiftVector[i] = static_cast<int>(indexShiftVectorTemp[i] + 0.5 * (indexShiftVectorTemp[i] > 0 ? 1 : -1));
   }
 
-  indexShift = (getMapFrameToBufferOrderTransformation() * indexShiftVector).array();
+  indexShift = transformMapFrameToBufferOrder(indexShiftVector);
   return true;
 }
 
@@ -173,7 +184,7 @@ bool getPositionShiftFromIndexShift(Vector& positionShift,
                                     const Index& indexShift,
                                     const double& resolution)
 {
-  positionShift = (getBufferOrderToMapFrameTransformation() * indexShift.matrix()).cast<double>() * resolution;
+  positionShift = transformBufferOrderToMapFrame(indexShift) * resolution;
   return true;
 }
 
@@ -206,10 +217,25 @@ void wrapIndexToRange(Index& index, const Size& bufferSize)
   }
 }
 
-void wrapIndexToRange(int& index, const int& bufferSize)
+void wrapIndexToRange(int& index, int bufferSize)
 {
-  if (index < 0) index += ((-index / bufferSize) + 1) * bufferSize;
-  index = index % bufferSize;
+  // Try shortcuts before resorting to the expensive modulo operation.
+  if (index < bufferSize){
+    if(index >= 0){ // within the wanted range
+      return;
+    } else if(index >= -bufferSize){ // Index is below range, but not more than one span of the range.
+      index +=bufferSize;
+      return;
+    }else{ // Index is largely below range.
+      index = index % bufferSize;
+      index += bufferSize;
+    }
+  }else if(index < bufferSize*2){ // Index is above range, but not more than one span of the range.
+    index -= bufferSize;
+    return;
+  } else{ // Index is largely above range.
+    index = index % bufferSize;
+  }
 }
 
 void boundPositionToRange(Position& position, const Length& mapLength, const Position& mapPosition)
@@ -256,16 +282,16 @@ bool getSubmapInformation(Index& submapTopLeftIndex,
                           const Index& bufferStartIndex)
 {
   // (Top left / bottom right corresponds to the position in the matrix, not the map frame)
-  Eigen::Matrix2d transform = getMapFrameToBufferOrderTransformation().cast<double>();
+  const Eigen::Matrix2d halfTransform = 0.5 * getMapFrameToBufferOrderTransformation().cast<double>();
 
   // Corners of submap.
-  Position topLeftPosition = requestedSubmapPosition - transform * 0.5 * requestedSubmapLength.matrix();
+  Position topLeftPosition = requestedSubmapPosition - halfTransform * requestedSubmapLength.matrix();
   boundPositionToRange(topLeftPosition, mapLength, mapPosition);
   if(!getIndexFromPosition(submapTopLeftIndex, topLeftPosition, mapLength, mapPosition, resolution, bufferSize, bufferStartIndex)) return false;
   Index topLeftIndex;
   topLeftIndex = getIndexFromBufferIndex(submapTopLeftIndex, bufferSize, bufferStartIndex);
 
-  Position bottomRightPosition = requestedSubmapPosition + transform * 0.5 * requestedSubmapLength.matrix();
+  Position bottomRightPosition = requestedSubmapPosition + halfTransform * requestedSubmapLength.matrix();
   boundPositionToRange(bottomRightPosition, mapLength, mapPosition);
   Index bottomRightIndex;
   if(!getIndexFromPosition(bottomRightIndex, bottomRightPosition, mapLength, mapPosition, resolution, bufferSize, bufferStartIndex)) return false;
@@ -274,7 +300,7 @@ bool getSubmapInformation(Index& submapTopLeftIndex,
   // Get the position of the top left corner of the generated submap.
   Position topLeftCorner;
   if(!getPositionFromIndex(topLeftCorner, submapTopLeftIndex, mapLength, mapPosition, resolution, bufferSize, bufferStartIndex)) return false;
-  topLeftCorner -= transform * Position::Constant(0.5 * resolution);
+  topLeftCorner -= halfTransform * Position::Constant(resolution);
 
   // Size of submap.
   submapBufferSize = bottomRightIndex - topLeftIndex + Index::Ones();
@@ -289,9 +315,7 @@ bool getSubmapInformation(Index& submapTopLeftIndex,
 
   // Get the index of the cell which corresponds the requested
   // position of the submap.
-  if(!getIndexFromPosition(requestedIndexInSubmap, requestedSubmapPosition, submapLength, submapPosition, resolution, submapBufferSize)) return false;
-
-  return true;
+  return getIndexFromPosition(requestedIndexInSubmap, requestedSubmapPosition, submapLength, submapPosition, resolution, submapBufferSize);
 }
 
 Size getSubmapSizeFromCornerIndeces(const Index& topLeftIndex, const Index& bottomRightIndex,
@@ -493,21 +517,6 @@ Index getIndexFromLinearIndex(const size_t linearIndex, const Size& bufferSize, 
   return Index((int)linearIndex / bufferSize(1), (int)linearIndex % bufferSize(1));
 }
 
-void getIndicesForRegion(const Index& regionIndex, const Size& regionSize,
-                         std::vector<Index> indices)
-{
-//  for (int i = line.index_; col < line.endIndex(); col++) {
-//    for (int i = 0; i < getSize()(0); i++) {
-//
-//    }
-//  }
-}
-
-void getIndicesForRegions(const std::vector<Index>& regionIndeces, const Size& regionSizes,
-                          std::vector<Index> indices)
-{
-}
-
 bool colorValueToVector(const unsigned long& colorValue, Eigen::Vector3i& colorVector)
 {
   colorVector(0) = (colorValue >> 16) & 0x0000ff;
@@ -540,9 +549,9 @@ bool colorVectorToValue(const Eigen::Vector3i& colorVector, unsigned long& color
 
 void colorVectorToValue(const Eigen::Vector3i& colorVector, float& colorValue)
 {
-  unsigned long color = (colorVector(0) << 16) + (colorVector(1) << 8) + colorVector(2);
-  // cppcheck-suppress invalidPointerCast
-  colorValue = *reinterpret_cast<float*>(&color);
+  Color colors;
+  colors.longColor = (colorVector(0) << 16) + (colorVector(1) << 8) + colorVector(2);
+  colorValue = colors.floatColor;
 }
 
 void colorVectorToValue(const Eigen::Vector3f& colorVector, float& colorValue)
