@@ -51,7 +51,7 @@ SpiralIterator& SpiralIterator::operator =(const SpiralIterator& other)
   return *this;
 }
 
-bool SpiralIterator::operator !=(const SpiralIterator& other) const
+bool SpiralIterator::operator !=(const SpiralIterator& /*other*/) const
 {
   return (pointsRing_.back() != pointsRing_.back()).any();
 }
@@ -101,16 +101,16 @@ void SpiralIterator::generateRing()
     normal.x() = -signum(point.y());
     normal.y() = signum(point.x());
     if (normal.x() != 0
-        && (int) Vector(point.x() + normal.x(), point.y()).norm() == distance_)
+        && static_cast<unsigned int>(Vector(point.x() + normal.x(), point.y()).norm()) == distance_)
       point.x() += normal.x();
     else if (normal.y() != 0
-        && (int) Vector(point.x(), point.y() + normal.y()).norm() == distance_)
+        && static_cast<unsigned int>(Vector(point.x(), point.y() + normal.y()).norm()) == distance_)
       point.y() += normal.y();
     else {
       point.x() += normal.x();
       point.y() += normal.y();
     }
-  } while (point.x() != distance_ || point.y() != 0);
+  } while (static_cast<unsigned int>(point.x()) != distance_ || point.y() != 0);
 }
 
 double SpiralIterator::getCurrentRadius() const
