@@ -100,22 +100,23 @@ class GridMapCvConverter
     grid_map::Matrix& data = gridMap[layer];
 
     for (GridMapIterator iterator(gridMap); !iterator.isPastEnd(); ++iterator) {
-      const Index index(*iterator);
+      const grid_map::Index gridMapIndex = *iterator;
+      const grid_map::Index imageIndex = iterator.getUnwrappedIndex();
 
       // Check for alpha layer.
       if (hasAlpha) {
-        const Type_ alpha = image.at<cv::Vec<Type_, NChannels_>>(index(0), index(1))[NChannels_ - 1];
+        const Type_ alpha = image.at<cv::Vec<Type_, NChannels_>>(imageIndex(0), imageIndex(1))[NChannels_ - 1];
         if (alpha < alphaTreshold) continue;
       }
 
       // Compute value.
-      const Type_ imageValue = imageMono.at<Type_>(index(0), index(1));
+      const Type_ imageValue = imageMono.at<Type_>(imageIndex(0), imageIndex(1));
       const float mapValue = lowerValue + mapValueDifference * ((float) imageValue / maxImageValue);
-      data(index(0), index(1)) = mapValue;
+      data(gridMapIndex(0), gridMapIndex(1)) = mapValue;
     }
 
     return true;
-  };
+  }
 
   /*!
    * Adds a color layer with data from an image.
@@ -176,7 +177,7 @@ class GridMapCvConverter
     const float minValue = gridMap.get(layer).minCoeffOfFinites();
     const float maxValue = gridMap.get(layer).maxCoeffOfFinites();
     return toImage<Type_, NChannels_>(gridMap, layer, encoding, minValue, maxValue, image);
-  };
+  }
 
   /*!
    * Creates a cv mat from a grid map layer.
@@ -225,9 +226,9 @@ class GridMapCvConverter
 
     for (GridMapIterator iterator(map); !iterator.isPastEnd(); ++iterator) {
       const Index index(*iterator);
-      if (std::isfinite(data(index(0), index(1)))) {
-        const float& value = data(index(0), index(1));
-        const Type_ imageValue = (Type_) (((value - lowerValue) / (upperValue - lowerValue)) * (float) imageMax);
+      const float& value = data(index(0), index(1));
+      if (std::isfinite(value)) {
+        const Type_ imageValue = (Type_)(((value - lowerValue) / (upperValue - lowerValue)) * (float)imageMax);
         const Index imageIndex(iterator.getUnwrappedIndex());
         unsigned int channel = 0;
         image.at<cv::Vec<Type_, NChannels_>>(imageIndex(0), imageIndex(1))[channel] = imageValue;
@@ -243,7 +244,7 @@ class GridMapCvConverter
     }
 
     return true;
-  };
+  }
 
 };
 
