@@ -11,7 +11,11 @@
 
 #include <grid_map_ros/grid_map_ros.hpp>
 
-#include <filters/filter_chain.h>
+#pragma GCC diagnostic push
+#pragma GCC diagnostic ignored "-Wpedantic"
+#pragma GCC diagnostic ignored "-Wformat"
+#include <filters/filter_chain.hpp>
+#pragma GCC diagnostic pop
 #include <ros/ros.h>
 #include <string>
 
