@@ -11,35 +11,34 @@
 #include <grid_map_core/grid_map_core.hpp>
 
 // PCL
-#include <pcl/point_cloud.h>
-#include <pcl/point_types.h>
-#include <pcl/conversions.h>
 #include <pcl/PolygonMesh.h>
 #include <pcl/common/common.h>
+#include <pcl/conversions.h>
+#include <pcl/point_cloud.h>
+#include <pcl/point_types.h>
 
 // STD
-#include <iostream>
-#include <vector>
 #include <algorithm>
 #include <cmath>
+#include <iostream>
+#include <vector>
 
 namespace grid_map {
 
 /*!
  * Conversions between grid maps and PCL types.
  */
-class GridMapPclConverter
-{
+class GridMapPclConverter {
  public:
   /*!
    * Default constructor.
    */
-  GridMapPclConverter();
+  GridMapPclConverter() = default;
 
   /*!
    * Destructor.
    */
-  virtual ~GridMapPclConverter();
+  virtual ~GridMapPclConverter() = default;
 
   /*!
    * Initializes the geometry of a grid map from a polygon mesh. This changes
@@ -49,8 +48,7 @@ class GridMapPclConverter
    * @param[out] gridMap the grid map to be initialized.
    * @return true if successful, false otherwise.
    */
-  static bool initializeFromPolygonMesh(const pcl::PolygonMesh& mesh, const double resolution,
-                                        grid_map::GridMap& gridMap);
+  static bool initializeFromPolygonMesh(const pcl::PolygonMesh& mesh, const double resolution, grid_map::GridMap& gridMap);
 
   /*!
    * Adds a layer with data from a polygon mesh. The mesh is ray traced from
@@ -60,15 +58,11 @@ class GridMapPclConverter
    * @param[out] gridMap the grid map to be populated.
    * @return true if successful, false otherwise.
    */
-  static bool addLayerFromPolygonMesh(const pcl::PolygonMesh& mesh, const std::string& layer,
-                                      grid_map::GridMap& gridMap);
+  static bool addLayerFromPolygonMesh(const pcl::PolygonMesh& mesh, const std::string& layer, grid_map::GridMap& gridMap);
 
  private:
-  static bool rayTriangleIntersect(const Eigen::Vector3f& point,
-                                   const Eigen::Vector3f& ray,
-                                   const Eigen::Matrix3f& triangleVertices,
+  static bool rayTriangleIntersect(const Eigen::Vector3f& point, const Eigen::Vector3f& ray, const Eigen::Matrix3f& triangleVertices,
                                    Eigen::Vector3f& intersectionPoint);
-
 };
 
-} /* namespace */
+} /* namespace grid_map */
