@@ -10,40 +10,25 @@
 
 namespace grid_map {
 
-GridMapPclConverter::GridMapPclConverter()
-{
-}
-
-GridMapPclConverter::~GridMapPclConverter()
-{
-}
-
-bool GridMapPclConverter::initializeFromPolygonMesh(const pcl::PolygonMesh& mesh,
-                                                    const double resolution,
-                                                    grid_map::GridMap& gridMap)
-{
-  pcl::PointCloud < pcl::PointXYZ > cloud;
+bool GridMapPclConverter::initializeFromPolygonMesh(const pcl::PolygonMesh& mesh, const double resolution, grid_map::GridMap& gridMap) {
+  pcl::PointCloud<pcl::PointXYZ> cloud;
   pcl::fromPCLPointCloud2(mesh.cloud, cloud);
   pcl::PointXYZ minBound;
   pcl::PointXYZ maxBound;
   pcl::getMinMax3D(cloud, minBound, maxBound);
 
   grid_map::Length length = grid_map::Length(maxBound.x - minBound.x, maxBound.y - minBound.y);
-  grid_map::Position position = grid_map::Position((maxBound.x + minBound.x) / 2.0,
-                                                   (maxBound.y + minBound.y) / 2.0);
+  grid_map::Position position = grid_map::Position((maxBound.x + minBound.x) / 2.0, (maxBound.y + minBound.y) / 2.0);
   gridMap.setGeometry(length, resolution, position);
 
   return true;
 }
 
-bool GridMapPclConverter::addLayerFromPolygonMesh(const pcl::PolygonMesh& mesh,
-                                                  const std::string& layer,
-                                                  grid_map::GridMap& gridMap)
-{
+bool GridMapPclConverter::addLayerFromPolygonMesh(const pcl::PolygonMesh& mesh, const std::string& layer, grid_map::GridMap& gridMap) {
   // Adding a layer to the grid map to put data into
   gridMap.add(layer);
   // Converting out of binary cloud data
-  pcl::PointCloud <pcl::PointXYZ> cloud;
+  pcl::PointCloud<pcl::PointXYZ> cloud;
   pcl::fromPCLPointCloud2(mesh.cloud, cloud);
   // Direction and max height for projection ray
   const Eigen::Vector3f ray = -Eigen::Vector3f::UnitZ();
@@ -71,15 +56,13 @@ bool GridMapPclConverter::addLayerFromPolygonMesh(const pcl::PolygonMesh& mesh,
     bool isSuccess;
     SubmapGeometry submap(gridMap, position, length, isSuccess);
     if (isSuccess) {
-      for (grid_map::SubmapIterator iterator(submap); !iterator.isPastEnd();
-           ++iterator) {
+      for (grid_map::SubmapIterator iterator(submap); !iterator.isPastEnd(); ++iterator) {
         // Cell position
         const Index index(*iterator);
         grid_map::Position vertexPositionXY;
         gridMap.getPosition(index, vertexPositionXY);
         // Ray origin
-        Eigen::Vector3f point(vertexPositionXY.x(), vertexPositionXY.y(),
-                              maxBound.z + 1.0);
+        Eigen::Vector3f point(vertexPositionXY.x(), vertexPositionXY.y(), maxBound.z + 1.0);
         // Vertical ray/triangle intersection
         Eigen::Vector3f intersectionPoint;
         if (rayTriangleIntersect(point, ray, triangleVertexMatrix, intersectionPoint)) {
@@ -97,10 +80,8 @@ bool GridMapPclConverter::addLayerFromPolygonMesh(const pcl::PolygonMesh& mesh,
   return true;
 }
 
-bool GridMapPclConverter::rayTriangleIntersect(
-    const Eigen::Vector3f& point, const Eigen::Vector3f& ray,
-    const Eigen::Matrix3f& triangleVertexMatrix,
-    Eigen::Vector3f& intersectionPoint) {
+bool GridMapPclConverter::rayTriangleIntersect(const Eigen::Vector3f& point, const Eigen::Vector3f& ray,
+                                               const Eigen::Matrix3f& triangleVertexMatrix, Eigen::Vector3f& intersectionPoint) {
   // Algorithm here is adapted from:
   // http://softsurfer.com/Archive/algorithm_0105/algorithm_0105.htm#intersect_RayTriangle()
   //
@@ -117,11 +98,15 @@ bool GridMapPclConverter::rayTriangleIntersect(
   const Eigen::Vector3f n = u.cross(v);
   const float n_dot_ray = n.dot(ray);
 
-  if (std::fabs(n_dot_ray) < 1e-9) return false;
+  if (std::fabs(n_dot_ray) < 1e-9) {
+    return false;
+  }
 
   const float r = n.dot(a - point) / n_dot_ray;
 
-  if (r < 0) return false;
+  if (r < 0) {
+    return false;
+  }
 
   // Note(alexmillane): The addition of this comparison delta (not in the
   // original algorithm) means that rays intersecting the edge of triangles are
@@ -132,15 +117,19 @@ bool GridMapPclConverter::rayTriangleIntersect(
   const float denominator = u.dot(v) * u.dot(v) - u.dot(u) * v.dot(v);
   const float s_numerator = u.dot(v) * w.dot(v) - v.dot(v) * w.dot(u);
   const float s = s_numerator / denominator;
-  if (s < (0 - delta) || s > (1 + delta)) return false;
+  if (s < (0 - delta) || s > (1 + delta)) {
+    return false;
+  }
 
   const float t_numerator = u.dot(v) * w.dot(u) - u.dot(u) * w.dot(v);
   const float t = t_numerator / denominator;
-  if (t < (0 - delta) || s + t > (1 + delta)) return false;
+  if (t < (0 - delta) || s + t > (1 + delta)) {
+    return false;
+  }
 
   intersectionPoint = a + s * u + t * v;
 
   return true;
 }
 
-} /* namespace */
+} /* namespace grid_map */
