@@ -19,7 +19,7 @@ class Vector3;
 class Quaternion;
 class ManualObject;
 class ColourValue;
-}
+}  // namespace Ogre
 
 namespace rviz {
 class BillboardLine;
@@ -28,19 +28,22 @@ class BillboardLine;
 namespace grid_map_rviz_plugin {
 
 // Visualizes a single grid_map_msgs::GridMap message.
-class GridMapVisual
-{
+class GridMapVisual {
  public:
+  using MaskArray = Eigen::Array<bool, Eigen::Dynamic, Eigen::Dynamic>;
+  using ColorArray = Eigen::Array<Ogre::ColourValue, Eigen::Dynamic, Eigen::Dynamic>;
+
   GridMapVisual(Ogre::SceneManager* sceneManager, Ogre::SceneNode* parentNode);
   virtual ~GridMapVisual();
 
   // Copy the grid map data to map_.
   void setMessage(const grid_map_msgs::GridMap::ConstPtr& msg);
   // Compute the visualization of map_.
-  void computeVisualization(float alpha, bool showGridLines, bool flatTerrain, std::string heightLayer, bool flatColor,
-                            bool noColor, Ogre::ColourValue meshColor, bool mapLayerColor, std::string colorLayer,
-                            bool useRainbow, bool invertRainbow, Ogre::ColourValue minColor, Ogre::ColourValue maxColor,
-                            bool autocomputeIntensity, float minIntensity, float maxIntensity);
+
+  void computeVisualization(float alpha, bool showGridLines, bool flatTerrain, std::string heightLayer, bool flatColor, bool noColor,
+                            Ogre::ColourValue meshColor, bool mapLayerColor, std::string colorLayer, bool useRainbow, bool invertRainbow,
+                            Ogre::ColourValue minColor, Ogre::ColourValue maxColor, bool autocomputeIntensity, float minIntensity,
+                            float maxIntensity);
 
   // Set the coordinate frame pose.
   void setFramePosition(const Ogre::Vector3& position);
@@ -50,6 +53,8 @@ class GridMapVisual
   std::vector<std::string> getLayerNames();
 
  private:
+  enum class ColoringMethod { FLAT, COLOR_LAYER, INTENSITY_LAYER_MANUAL, INTENSITY_LAYER_RAINBOW, INTENSITY_LAYER_INVERTED_RAINBOW };
+
   Ogre::SceneNode* frameNode_;
   Ogre::SceneManager* sceneManager_;
 
@@ -63,13 +68,74 @@ class GridMapVisual
 
   // Grid map.
   grid_map::GridMap map_;
-  bool haveMap_;
 
   // Helper methods.
-  void normalizeIntensity(float& intensity, float minIntensity, float maxIntensity);
-  Ogre::ColourValue getRainbowColor(float intensity);
-  Ogre::ColourValue getInterpolatedColor(float intensity, Ogre::ColourValue minColor,
-                                         Ogre::ColourValue maxColor);
+  bool haveMap_;
+  /**
+   * Initialized the manualObject if not already initialized and clears all previously added data.
+   * @param nVertices An estimate on the number of vertices to be added.
+   */
+  void initializeAndBeginManualObject(size_t nVertices);
+
+  /**
+   * Computes a matrix of color values corresponding to the grid cells. Color is computed depending on the coloringMethod.
+   * @param heightData Height values of the cells.
+   * @param colorData Values of the layer specified for coloring the mesh.
+   * @param coloringMethod The strategy to color, see ColoringMethod.
+   * @param flatColor Used only if coloringMethod is FLAT
+   * @param minIntensity Used for the intensity based coloring methods only.
+   * @param maxIntensity Used for the intensity based coloring methods only.
+   * @param autocomputeIntensity Wheter to override the values in min and max intensity and compute them based on the provided intensity
+   * data.
+   * @param minColor Used only if coloringMethod is COLOR_LAYER.
+   * @param maxColor Used only if coloringMethod is COLOR_LAYER.
+   * @return The color for each cell.
+   */
+  ColorArray computeColorValues(Eigen::Ref<const grid_map::Matrix> heightData, Eigen::Ref<const grid_map::Matrix> colorData,
+                                GridMapVisual::ColoringMethod coloringMethod, Ogre::ColourValue flatColor, double minIntensity,
+                                double maxIntensity, bool autocomputeIntensity, Ogre::ColourValue minColor, Ogre::ColourValue maxColor);
+
+  /**
+   * Initialized the meshLines_ object. Should be called before adding lines. Sets the drawing style and allocates the buffer.
+   * @param cols Number of columns that will be drawn.
+   * @param rows Number of rows that will be drawn.
+   * @param resolution Resolution of the map. Used to compute the line thickness.
+   * @param alpha Line opacity.
+   */
+  void initializeMeshLines(size_t cols, size_t rows, double resolution, double alpha);
+
+  /**
+   * Computes a mask where all the provided basicLayers are finite. Used to do fast lockups during mesh creation.
+   * @param basicLayers The values to check, if they are finite.
+   * @return A boolean mask. True indicates that all layers are finite, false indicates that at least one layer is NAN.
+   */
+  MaskArray computeIsValidMask(std::vector<std::string> basicLayers);
+
+  /**
+   * Transforms the intensity into [0,1] range where 0 corresponds to the minIntensity and 1 to maxIntensity. The given value is clipped to
+   * that range.
+   * @param intensity The intensity value to normalize.
+   * @param minIntensity Lower bound.
+   * @param maxIntensity Upper bound.
+   */
+  static void normalizeIntensity(float& intensity, float minIntensity, float maxIntensity);
+
+  /**
+   * Copied from rviz/src/rviz/default_plugin/point_cloud_transformers.cpp. Transforms an intensity value in [0,1] range to a rainbow
+   * coloring.
+   * @param intensity Value to color, should be in [0,1] range, otherwise it is clipped.
+   * @return The corresponding rainbow color.
+   */
+  static Ogre::ColourValue getRainbowColor(float intensity);
+
+  /**
+   * Returns a linearly interpolated color between min and max Color with the weighting [0,1] specified by intensity.
+   * @param intensity The weighting for interpolation, 0 means minColor, 1 means maxColor.
+   * @param minColor The lower color for interpolation.
+   * @param maxColor The upper color for interpolation.
+   * @return
+   */
+  Ogre::ColourValue getInterpolatedColor(float intensity, Ogre::ColourValue minColor, Ogre::ColourValue maxColor);
 };
 
-}  // namespace
+}  // namespace grid_map_rviz_plugin
