@@ -6,28 +6,25 @@
  *  Institute: ETH Zurich, ANYbotics
  */
 
-#include <rviz/uniform_string_stream.h>
-#include <OGRE/OgreMaterialManager.h>
-#include <OGRE/OgreTextureManager.h>
-#include <OGRE/OgreTechnique.h>
-
-#include <OGRE/OgreVector3.h>
-#include <OGRE/OgreSceneNode.h>
-#include <OGRE/OgreSceneManager.h>
-#include <OGRE/OgreManualObject.h>
-
-#include <rviz/ogre_helpers/billboard_line.h>
-
-#include <grid_map_ros/grid_map_ros.hpp>
-#include <grid_map_core/GridMapMath.hpp>
 #include "grid_map_rviz_plugin/GridMapVisual.hpp"
+
+#include <OGRE/OgreManualObject.h>
+#include <OGRE/OgreMaterialManager.h>
+#include <OGRE/OgreSceneManager.h>
+#include <OGRE/OgreSceneNode.h>
+#include <OGRE/OgreTechnique.h>
+#include <OGRE/OgreTextureManager.h>
+#include <OGRE/OgreVector3.h>
+#include <rviz/ogre_helpers/billboard_line.h>
+#include <rviz/uniform_string_stream.h>
+#include <chrono>
+
+#include <grid_map_core/GridMapMath.hpp>
+#include <grid_map_ros/grid_map_ros.hpp>
 
 namespace grid_map_rviz_plugin {
 
-GridMapVisual::GridMapVisual(Ogre::SceneManager* sceneManager, Ogre::SceneNode* parentNode)
-    : manualObject_(0),
-      haveMap_(false)
-{
+GridMapVisual::GridMapVisual(Ogre::SceneManager* sceneManager, Ogre::SceneNode* parentNode) : manualObject_(0), haveMap_(false) {
   sceneManager_ = sceneManager;
   frameNode_ = parentNode->createChildSceneNode();
 
@@ -35,8 +32,7 @@ GridMapVisual::GridMapVisual(Ogre::SceneManager* sceneManager, Ogre::SceneNode* 
   meshLines_.reset(new rviz::BillboardLine(sceneManager_, frameNode_));
 }
 
-GridMapVisual::~GridMapVisual()
-{
+GridMapVisual::~GridMapVisual() {
   // Destroy the ManualObject.
   sceneManager_->destroyManualObject(manualObject_);
   material_->unload();
@@ -46,26 +42,24 @@ GridMapVisual::~GridMapVisual()
   sceneManager_->destroySceneNode(frameNode_);
 }
 
-void GridMapVisual::setMessage(const grid_map_msgs::GridMap::ConstPtr& msg)
-{
+void GridMapVisual::setMessage(const grid_map_msgs::GridMap::ConstPtr& msg) {
   // Convert grid map message.
   grid_map::GridMapRosConverter::fromMessage(*msg, map_);
   haveMap_ = true;
 }
 
-void GridMapVisual::computeVisualization(float alpha, bool showGridLines, bool flatTerrain, std::string heightLayer,
-                                         bool flatColor, bool noColor, Ogre::ColourValue meshColor, bool mapLayerColor,
-                                         std::string colorLayer, bool useRainbow, bool invertRainbow,
-                                         Ogre::ColourValue minColor, Ogre::ColourValue maxColor,
-                                         bool autocomputeIntensity, float minIntensity, float maxIntensity)
-{
+void GridMapVisual::computeVisualization(float alpha, bool showGridLines, bool flatTerrain, std::string heightLayer, bool flatColor,
+                                         bool noColor, Ogre::ColourValue meshColor, bool mapLayerColor, std::string colorLayer,
+                                         bool useRainbow, bool invertRainbow, Ogre::ColourValue minColor, Ogre::ColourValue maxColor,
+                                         bool autocomputeIntensity, float minIntensity, float maxIntensity) {
+  const auto startTime = std::chrono::high_resolution_clock::now();
   if (!haveMap_) {
     ROS_DEBUG("Unable to visualize grid map, no map data. Use setMessage() first!");
     return;
   }
 
   // Get list of layers and check if the requested ones are present.
-  std::vector<std::string> layerNames = map_.getLayers();
+  const std::vector<std::string> layerNames = map_.getLayers();
   if (layerNames.size() < 1) {
     ROS_DEBUG("Unable to visualize grid map, map must contain at least one layer.");
     return;
@@ -89,125 +83,131 @@ void GridMapVisual::computeVisualization(float alpha, bool showGridLines, bool f
   const grid_map::Matrix& heightData = map_[flatTerrain ? layerNames[0] : heightLayer];
   const grid_map::Matrix& colorData = map_[flatColor ? layerNames[0] : colorLayer];
 
-  // initialize ManualObject
-  if (!manualObject_) {
-    static uint32_t count = 0;
-    rviz::UniformStringStream ss;
-    ss << "Mesh" << count++;
-    manualObject_ = sceneManager_->createManualObject(ss.str());
-    frameNode_->attachObject(manualObject_);
+  // Reset and begin the manualObject (mesh).
+  // For more information: https://www.ogre3d.org/docs/api/1.7/class_ogre_1_1_manual_object.html#details
+  const size_t nVertices = cols * rows;
+  initializeAndBeginManualObject(nVertices);
 
-    ss << "Material";
-    materialName_ = ss.str();
-    material_ = Ogre::MaterialManager::getSingleton().create(materialName_, "rviz");
-    material_->setReceiveShadows(false);
-    material_->getTechnique(0)->setLightingEnabled(true);
-    material_->setCullingMode(Ogre::CULL_NONE);
-  }
-
-  manualObject_->clear();
-  size_t nVertices = 4 + 6 * (cols * rows - cols - rows);
-  manualObject_->estimateVertexCount(nVertices);
-  manualObject_->begin(materialName_, Ogre::RenderOperation::OT_TRIANGLE_LIST);
-
+  // Reset the mesh lines.
   meshLines_->clear();
   if (showGridLines) {
-    meshLines_->setColor(0.0, 0.0, 0.0, alpha);
-    meshLines_->setLineWidth(resolution / 10.0);
-    meshLines_->setMaxPointsPerLine(2);
-    // In the algorithm below, we have to account for max. 4 lines per cell.
-    size_t nLines = 2 * (rows * (cols - 1) + cols * (rows - 1));
-    meshLines_->setNumLines(nLines);
+    initializeMeshLines(cols, rows, resolution, alpha);
   }
 
-  // Determine max and min intensity.
-  if (autocomputeIntensity && !flatColor && !mapLayerColor) {
-    minIntensity = colorData.minCoeffOfFinites();
-    maxIntensity = colorData.maxCoeffOfFinites();
+  // Compute a mask of valid cells.
+  auto basicLayers = map_.getBasicLayers();
+  if (std::find(basicLayers.begin(), basicLayers.end(), heightLayer) == basicLayers.end()) {
+    basicLayers.emplace_back(heightLayer);
+  }
+  const MaskArray isValid = computeIsValidMask(basicLayers);
+
+  // Compute the display heights for each cell.
+  Eigen::ArrayXXf heightOrFlatData;
+  if (flatTerrain) {
+    heightOrFlatData = Eigen::ArrayXXf::Zero(heightData.rows(), heightData.cols());
+  } else {
+    heightOrFlatData = heightData.array();
   }
 
-  if (!map_.hasBasicLayers()) map_.setBasicLayers({heightLayer});
+  // Compute the color data for each cell.
+  ColoringMethod coloringMethod;
+  if (flatColor) {
+    coloringMethod = ColoringMethod::FLAT;
+  } else if(mapLayerColor) {
+    coloringMethod = ColoringMethod::COLOR_LAYER;
+  } else if (!useRainbow) {
+    coloringMethod = ColoringMethod::INTENSITY_LAYER_MANUAL;
+  } else if (!invertRainbow) {
+    coloringMethod = ColoringMethod::INTENSITY_LAYER_RAINBOW;
+  } else {
+    coloringMethod = ColoringMethod::INTENSITY_LAYER_INVERTED_RAINBOW;
+  }
 
-  // Plot mesh.
-  for (size_t i = 0; i < rows - 1; ++i) {
-    for (size_t j = 0; j < cols - 1; ++j) {
-      std::vector<Ogre::Vector3> vertices;
+  const auto colorValues = computeColorValues(heightData, colorData, coloringMethod, meshColor, minIntensity, maxIntensity,
+                                              autocomputeIntensity, minColor, maxColor);
+
+  // Initialize loop constants.
+  grid_map::Position topLeft;
+  map_.getPosition(grid_map::Index(0, 0), topLeft);
+
+  Eigen::ArrayXXi indexToOgreIndex;
+  indexToOgreIndex.setConstant(rows, cols, -1);
+
+  int ogreIndex = 0;
+
+  // Add vertices for mesh.
+  for (size_t i = 0; i < rows; ++i) {
+    for (size_t j = 0; j < cols; ++j) {
+      std::vector<int> vertices;
       std::vector<Ogre::ColourValue> colors;
+
+      // Add the vertex to the scene
+      grid_map::Index index(i, j);
+      if (!isValid(index(0), index(1))) {
+        continue;
+      }
+      grid_map::Position position = topLeft.array() - index.cast<double>() * resolution;
+      manualObject_->position(position(0), position(1), heightOrFlatData(index(0), index(1)));
+
+      const Ogre::ColourValue& color = colorValues(index(0), index(1));
+      manualObject_->colour(color.r, color.g, color.b, alpha);
+
+      indexToOgreIndex(index(0), index(1)) = ogreIndex;
+      ogreIndex++;
+
+      // We can only add triangles to the top left side of the current vertex if we have data.
+      if (i == 0 || j == 0) {
+        continue;
+      }
+
+      // Add triangles and grid to scene.
+      std::vector<int> vertexIndices;
+      std::vector<Ogre::Vector3> vertexPositions;
       for (size_t k = 0; k < 2; k++) {
         for (size_t l = 0; l < 2; l++) {
-          grid_map::Position position;
-          grid_map::Index index(i + k, j + l);
-          if (!map_.isValid(index)) continue;
-
-          map_.getPosition(index, position);
-          float height = heightData(index(0), index(1));
-          vertices.push_back(Ogre::Vector3(position(0), position(1), flatTerrain ? 0.0 : height));
-          if (!flatColor) {
-            float color = colorData(index(0), index(1));
-            Ogre::ColourValue colorValue;
-            if (mapLayerColor) {
-              Eigen::Vector3f colorVectorRGB;
-              grid_map::colorValueToVector(color, colorVectorRGB);
-              colorValue = Ogre::ColourValue(colorVectorRGB(0), colorVectorRGB(1), colorVectorRGB(2));
-            } else {
-              normalizeIntensity(color, minIntensity, maxIntensity);
-              colorValue = useRainbow ? (invertRainbow ? getRainbowColor(1.0f - color)
-                                                       : getRainbowColor(color))
-                                      : getInterpolatedColor(color, minColor, maxColor);
-            }
-            colors.push_back(colorValue);
+          grid_map::Index index(i - k, j - l);
+          if (!isValid(index(0), index(1))) {
+            continue;
           }
+          const grid_map::Position position = topLeft.array() - index.cast<double>() * resolution;
+          vertexIndices.emplace_back(indexToOgreIndex(index(0), index(1)));
+          vertexPositions.emplace_back(position(0), position(1), heightOrFlatData(index(0), index(1)));
         }
       }
 
       // Plot triangles if we have enough vertices.
-      if (vertices.size() > 2) {
-        Ogre::Vector3 normal = vertices.size() == 4
-                               ? (vertices[3] - vertices[0]).crossProduct(vertices[2] -vertices[1])
-                               : (vertices[2] - vertices[1]).crossProduct(vertices[1] - vertices[0]);
-        normal.normalise();
+      if (vertexIndices.size() > 2) {
         // Create one or two triangles from the vertices depending on how many vertices we have.
         if (!noColor) {
-          for (size_t m = 1; m < vertices.size() - 1; m++) {
-            manualObject_->position(vertices[m-1]);
-            manualObject_->normal(normal);
-            Ogre::ColourValue color = flatColor ? meshColor : colors[m-1];
-            manualObject_->colour(color.r, color.g, color.b, alpha);
-
-            manualObject_->position(vertices[m]);
-            manualObject_->normal(normal);
-            color = flatColor ? meshColor : colors[m];
-            manualObject_->colour(color.r, color.g, color.b, alpha);
-
-            manualObject_->position(vertices[m+1]);
-            manualObject_->normal(normal);
-            color = flatColor ? meshColor : colors[m+1];
-            manualObject_->colour(color.r, color.g, color.b, alpha);
+          if (vertexIndices.size() == 3) {
+            manualObject_->triangle(vertexIndices[0], vertexIndices[1], vertexIndices[2]);
+          } else {
+            manualObject_->quad(vertexIndices[0], vertexIndices[2], vertexIndices[3], vertexIndices[1]);
           }
         }
 
         // plot grid lines
         if (showGridLines) {
-          meshLines_->addPoint(vertices[0]);
-          meshLines_->addPoint(vertices[1]);
+          meshLines_->addPoint(vertexPositions[0]);
+          meshLines_->addPoint(vertexPositions[1]);
           meshLines_->newLine();
 
-          if (vertices.size() == 3) {
-            meshLines_->addPoint(vertices[1]);
-            meshLines_->addPoint(vertices[2]);
+          if (vertexIndices.size() == 3) {
+            meshLines_->addPoint(vertexPositions[1]);
+            meshLines_->addPoint(vertexPositions[2]);
             meshLines_->newLine();
           } else {
-            meshLines_->addPoint(vertices[1]);
-            meshLines_->addPoint(vertices[3]);
+            meshLines_->addPoint(vertexPositions[1]);
+            meshLines_->addPoint(vertexPositions[3]);
             meshLines_->newLine();
 
-            meshLines_->addPoint(vertices[3]);
-            meshLines_->addPoint(vertices[2]);
+            meshLines_->addPoint(vertexPositions[3]);
+            meshLines_->addPoint(vertexPositions[2]);
             meshLines_->newLine();
           }
 
-          meshLines_->addPoint(vertices[2]);
-          meshLines_->addPoint(vertices[0]);
+          meshLines_->addPoint(vertexPositions[2]);
+          meshLines_->addPoint(vertexPositions[0]);
           meshLines_->newLine();
         }
       }
@@ -224,34 +224,114 @@ void GridMapVisual::computeVisualization(float alpha, bool showGridLines, bool f
     material_->getTechnique(0)->setSceneBlending(Ogre::SBT_REPLACE);
     material_->getTechnique(0)->setDepthWriteEnabled(true);
   }
+
+  const auto stopTime = std::chrono::high_resolution_clock::now();
+  const auto elapsedTime = std::chrono::duration_cast<std::chrono::milliseconds>(stopTime - startTime);
+  ROS_DEBUG_STREAM("Visualization of grid_map took: " << elapsedTime.count() << " ms.");
 }
 
-void GridMapVisual::setFramePosition(const Ogre::Vector3& position)
-{
+void GridMapVisual::initializeAndBeginManualObject(size_t nVertices) {
+  if (!manualObject_) {
+    static uint32_t count = 0;
+    rviz::UniformStringStream ss;
+    ss << "Mesh" << count++;
+    manualObject_ = sceneManager_->createManualObject(ss.str());
+    manualObject_->setDynamic(true);
+    frameNode_->attachObject(manualObject_);
+
+    ss << "Material";
+    materialName_ = ss.str();
+    material_ = Ogre::MaterialManager::getSingleton().create(materialName_, Ogre::ResourceGroupManager::DEFAULT_RESOURCE_GROUP_NAME);
+    material_->setReceiveShadows(false);
+    material_->getTechnique(0)->setLightingEnabled(true);
+    material_->setCullingMode(Ogre::CULL_NONE);
+  }
+
+  manualObject_->clear();
+  manualObject_->estimateVertexCount(nVertices);
+  manualObject_->begin(materialName_, Ogre::RenderOperation::OT_TRIANGLE_LIST);
+}
+
+GridMapVisual::ColorArray GridMapVisual::computeColorValues(Eigen::Ref<const grid_map::Matrix> heightData,
+                                                            Eigen::Ref<const grid_map::Matrix> colorData,
+                                                            GridMapVisual::ColoringMethod coloringMethod, Ogre::ColourValue flatColor,
+                                                            double minIntensity, double maxIntensity, bool autocomputeIntensity,
+                                                            Ogre::ColourValue minColor, Ogre::ColourValue maxColor) {
+  // Determine max and min intensity.
+  bool isIntensityColoringMethod = coloringMethod == ColoringMethod::INTENSITY_LAYER_INVERTED_RAINBOW ||
+                                   coloringMethod == ColoringMethod::INTENSITY_LAYER_RAINBOW ||
+                                   coloringMethod == ColoringMethod::INTENSITY_LAYER_MANUAL;
+  if (autocomputeIntensity && isIntensityColoringMethod) {
+    minIntensity = colorData.minCoeffOfFinites();
+    maxIntensity = minIntensity + std::max(colorData.maxCoeffOfFinites() - minIntensity, 0.2);  // regularize the intensity range.
+  }
+
+  switch (coloringMethod) {
+    case ColoringMethod::FLAT:
+      return ColorArray::Constant(heightData.rows(), heightData.cols(), flatColor);
+    case ColoringMethod::COLOR_LAYER:
+      return colorData.unaryExpr([](float color) {
+        Eigen::Vector3f colorVectorRGB;
+        grid_map::colorValueToVector(color, colorVectorRGB);
+        return Ogre::ColourValue(colorVectorRGB(0), colorVectorRGB(1), colorVectorRGB(2));
+      });
+    case ColoringMethod::INTENSITY_LAYER_MANUAL:
+      return colorData.unaryExpr([&](float color) {
+        normalizeIntensity(color, minIntensity, maxIntensity);
+        return getInterpolatedColor(color, minColor, maxColor);
+      });
+    case ColoringMethod::INTENSITY_LAYER_RAINBOW:
+      return colorData.unaryExpr([&](float color) {
+        normalizeIntensity(color, minIntensity, maxIntensity);
+        return getRainbowColor(color);
+      });
+    case ColoringMethod::INTENSITY_LAYER_INVERTED_RAINBOW:
+      return colorData.unaryExpr([&](float color) {
+        normalizeIntensity(color, minIntensity, maxIntensity);
+        return getRainbowColor(1.0f - color);
+      });
+    default:
+      throw std::invalid_argument(std::string("An unknown coloring method was provided: ") +
+                                  std::to_string(static_cast<int>(coloringMethod)));
+  }
+}
+
+void GridMapVisual::initializeMeshLines(size_t cols, size_t rows, double resolution, double alpha) {
+  meshLines_->setColor(0.0, 0.0, 0.0, alpha);
+  meshLines_->setLineWidth(resolution / 10.0);
+  meshLines_->setMaxPointsPerLine(2);
+  // In the algorithm below, we have to account for max. 4 lines per cell.
+  const size_t nLines = 2 * (rows * (cols - 1) + cols * (rows - 1));
+  meshLines_->setNumLines(nLines);
+}
+
+GridMapVisual::MaskArray GridMapVisual::computeIsValidMask(std::vector<std::string> basicLayers) {
+  MaskArray isValid = MaskArray::Ones(map_.getSize()(0), map_.getSize()(1));
+  for (const std::string& layer : basicLayers) {
+    isValid = isValid && map_.get(layer).array().unaryExpr([](float v) { return std::isfinite(v); });
+  }
+  return isValid;
+}
+
+void GridMapVisual::setFramePosition(const Ogre::Vector3& position) {
   frameNode_->setPosition(position);
 }
 
-void GridMapVisual::setFrameOrientation(const Ogre::Quaternion& orientation)
-{
+void GridMapVisual::setFrameOrientation(const Ogre::Quaternion& orientation) {
   frameNode_->setOrientation(orientation);
 }
 
-std::vector<std::string> GridMapVisual::getLayerNames()
-{
+std::vector<std::string> GridMapVisual::getLayerNames() {
   return map_.getLayers();
 }
 
-// Compute intensity value in the interval [0,1].
-void GridMapVisual::normalizeIntensity(float& intensity, float min_intensity, float max_intensity)
-{
+void GridMapVisual::normalizeIntensity(float& intensity, float min_intensity, float max_intensity) {
   intensity = std::min(intensity, max_intensity);
   intensity = std::max(intensity, min_intensity);
   intensity = (intensity - min_intensity) / (max_intensity - min_intensity);
 }
 
-// Copied from rviz/src/rviz/default_plugin/point_cloud_transformers.cpp.
-Ogre::ColourValue GridMapVisual::getRainbowColor(float intensity)
-{
+Ogre::ColourValue GridMapVisual::getRainbowColor(float intensity) {
   intensity = std::min(intensity, 1.0f);
   intensity = std::max(intensity, 0.0f);
 
@@ -262,19 +342,22 @@ Ogre::ColourValue GridMapVisual::getRainbowColor(float intensity)
   float n = 1 - f;
 
   Ogre::ColourValue color;
-  if (i <= 1) color[0] = n, color[1] = 0, color[2] = 1;
-  else if (i == 2) color[0] = 0, color[1] = n, color[2] = 1;
-  else if (i == 3) color[0] = 0, color[1] = 1, color[2] = n;
-  else if (i == 4) color[0] = n, color[1] = 1, color[2] = 0;
-  else if (i >= 5) color[0] = 1, color[1] = n, color[2] = 0;
+  if (i <= 1)
+    color[0] = n, color[1] = 0, color[2] = 1;
+  else if (i == 2)
+    color[0] = 0, color[1] = n, color[2] = 1;
+  else if (i == 3)
+    color[0] = 0, color[1] = 1, color[2] = n;
+  else if (i == 4)
+    color[0] = n, color[1] = 1, color[2] = 0;
+  else if (i >= 5)
+    color[0] = 1, color[1] = n, color[2] = 0;
 
   return color;
 }
 
 // Get interpolated color value.
-Ogre::ColourValue GridMapVisual::getInterpolatedColor(float intensity, Ogre::ColourValue min_color,
-                                                      Ogre::ColourValue max_color)
-{
+Ogre::ColourValue GridMapVisual::getInterpolatedColor(float intensity, Ogre::ColourValue min_color, Ogre::ColourValue max_color) {
   intensity = std::min(intensity, 1.0f);
   intensity = std::max(intensity, 0.0f);
 
@@ -286,4 +369,4 @@ Ogre::ColourValue GridMapVisual::getInterpolatedColor(float intensity, Ogre::Col
   return color;
 }
 
-}  // namespace
+}  // namespace grid_map_rviz_plugin
