@@ -32,7 +32,7 @@ namespace distance_transform
   const double GREEN_WEIGHT = 0.584;
   const double BLUE_WEIGHT = 0.114;
 
-  static image<uchar> *imageRGBtoGRAY(image<rgb> *input) {
+  static inline image<uchar> *imageRGBtoGRAY(image<rgb> *input) {
     int width = input->width();
     int height = input->height();
     image<uchar> *output = new image<uchar>(width, height, false);
@@ -48,7 +48,7 @@ namespace distance_transform
     return output;
   }
 
-  static image<rgb> *imageGRAYtoRGB(image<uchar> *input) {
+  static inline image<rgb> *imageGRAYtoRGB(image<uchar> *input) {
     int width = input->width();
     int height = input->height();
     image<rgb> *output = new image<rgb>(width, height, false);
@@ -63,7 +63,7 @@ namespace distance_transform
     return output;  
   }
 
-  static image<float> *imageUCHARtoFLOAT(image<uchar> *input) {
+  static inline image<float> *imageUCHARtoFLOAT(image<uchar> *input) {
     int width = input->width();
     int height = input->height();
     image<float> *output = new image<float>(width, height, false);
@@ -76,7 +76,7 @@ namespace distance_transform
     return output;  
   }
 
-  static image<float> *imageINTtoFLOAT(image<int> *input) {
+  static inline image<float> *imageINTtoFLOAT(image<int> *input) {
     int width = input->width();
     int height = input->height();
     image<float> *output = new image<float>(width, height, false);
@@ -108,13 +108,13 @@ namespace distance_transform
     return output;
   }
 
-  static image<uchar> *imageFLOATtoUCHAR(image<float> *input) {
+  static inline image<uchar> *imageFLOATtoUCHAR(image<float> *input) {
     float min, max;
     min_max(input, &min, &max);
     return imageFLOATtoUCHAR(input, min, max);
   }
 
-  static image<long> *imageUCHARtoLONG(image<uchar> *input) {
+  static inline image<long> *imageUCHARtoLONG(image<uchar> *input) {
     int width = input->width();
     int height = input->height();
     image<long> *output = new image<long>(width, height, false);
@@ -145,7 +145,7 @@ namespace distance_transform
     return output;
   }
 
-  static image<uchar> *imageLONGtoUCHAR(image<long> *input) {
+  static inline image<uchar> *imageLONGtoUCHAR(image<long> *input) {
     long min, max;
     min_max(input, &min, &max);
     return imageLONGtoUCHAR(input, min, max);
@@ -170,7 +170,7 @@ namespace distance_transform
     return output;
   }
 
-  static image<uchar> *imageSHORTtoUCHAR(image<short> *input) {
+  static inline image<uchar> *imageSHORTtoUCHAR(image<short> *input) {
     short min, max;
     min_max(input, &min, &max);
     return imageSHORTtoUCHAR(input, min, max);
