@@ -33,13 +33,13 @@ namespace distance_transform
   }
 
   template <class T>
-  inline T abs(const T &x) { return (x > 0 ? x : -x); };
+  inline T abs(const T &x) { return (x > 0 ? x : -x); }
 
   template <class T>
-  inline int sign(const T &x) { return (x >= 0 ? 1 : -1); };
+  inline int sign(const T &x) { return (x >= 0 ? 1 : -1); }
 
   template <class T>
-  inline T square(const T &x) { return x*x; };
+  inline T square(const T &x) { return x*x; }
 
   template <class T>
   inline T bound(const T &x, const T &min, const T &max) {
